refactor: move SpecListEntry to the tap types module
