+++ packages/app/src/tap/commands/specs-list.ts
@@ -1,11 +1,6 @@
 import type { FoundSpec } from '@packages/types'
 
-export interface SpecListEntry {
-  /** Project-relative spec path — the form `cypress run --spec` accepts. */
-  relativePath: string
-  /** Whether the spec is an end-to-end (integration) or component spec. */
-  specType: FoundSpec['specType']
-}
+import type { SpecListEntry } from '../types'
 
 export const toSpecListEntry = ({ relative, specType }: FoundSpec): SpecListEntry => {
   return { relativePath: relative, specType }
