+++ packages/app/src/tap/commands/specs.ts
@@ -1,7 +1,7 @@
 import { tapManagerDataSource } from '../TapManagerDataSource'
 import { defineCommand } from './definition'
 import { toSpecListEntry } from './specs-list'
-import type { SpecListEntry } from './specs-list'
+import type { SpecListEntry } from '../types'
 
 export const specsCommand = defineCommand({
   description: 'List all runnable specs for the selected Cypress instance.',
